Move some functions for reusability
