+++ applications/d2l-capture-central/src/components/my-videos/content-list.js
@@ -9,74 +9,17 @@ import '../relative-date.js';
 import './content-list-header.js';
 import './content-list-item-ghost.js';
 import './content-list-item.js';
+import { CaptureCentralList } from'../capture-central-list.js';
 
-import { bodyCompactStyles, bodySmallStyles } from '@brightspace-ui/core/components/typography/styles.js';
-import { css, html, LitElement } from 'lit-element/lit-element.js';
+import { html } from 'lit-element/lit-element.js';
 import { observe, toJS } from 'mobx';
 
-import { contentSearchMixin } from '../../mixins/content-search-mixin.js';
-import { DependencyRequester } from '../../mixins/dependency-requester-mixin.js';
-import { InternalLocalizeMixin } from '../../mixins/internal-localize-mixin.js';
-import { NavigationMixin } from '../../mixins/navigation-mixin.js';
-import { navigationSharedStyle } from '../../style/d2l-navigation-shared-styles.js';
 import { rootStore } from '../../state/root-store.js';
 
-class ContentList extends DependencyRequester(InternalLocalizeMixin(NavigationMixin(contentSearchMixin(LitElement)))) {
-	static get properties() {
-		return {
-			loading: { type: Boolean, attribute: false }
-		};
-	}
-
-	static get styles() {
-		return [bodyCompactStyles, bodySmallStyles, navigationSharedStyle, css`
-			:host([hidden]) {
-				display: none;
-			}
-
-			.title {
-				word-break: break-word;
-			}
-
-			#d2l-content-store-list {
-				padding-top: 1px;
-			}
-
-			.d2l-capture-central-content-list-no-results {
-				display: flex;
-				justify-content: center;
-				margin-top: 20px;
-			}
-
-			d2l-icon {
-				border-radius: 5px;
-				padding: 12px;
-				background-color: var(--d2l-color-celestine);
-				color: var(--d2l-color-white);
-			}
-		`];
-	}
-
+class ContentList extends CaptureCentralList {
 	constructor() {
 		super();
-		this.infiniteScrollThreshold = 400;
-		this.dateField = 'updatedAt';
-		this.loading = false;
-		this.alertToastMessage = '';
-		this.alertToastButtonText = '';
-		this.undoDeleteObject = {};
-
-		const {
-			searchQuery = '',
-			sortQuery = 'updatedAt:desc',
-			dateCreated = '',
-			dateModified = ''
-		} = rootStore.routingStore.getQueryParams();
-
-		this.queryParams = { searchQuery, sortQuery, dateCreated, dateModified };
-
-		window.addEventListener('scroll', this.onWindowScroll.bind(this));
-		this.observeQueryParams();
+		this.page = '/my-videos';
 	}
 
 	connectedCallback() {
@@ -111,37 +54,10 @@ class ContentList extends DependencyRequester(InternalLocalizeMixin(NavigationMi
 		`;
 	}
 
-	async addNewItemIntoContentItems(item) {
-		if (!item || !item.revision || !item.content || !item.upload) {
-			return;
-		}
-
-		const { title } = item.content;
-		const { id: revisionId, type } = item.revision;
-		const { id } = item.content;
-		const updatedAt = (new Date()).toISOString();
-
-		await this.insertIntoContentItemsBasedOnSort({
-			id, revisionId, type, title, updatedAt
-		});
-	}
-
 	areAnyFiltersActive() {
 		return this.queryParams.searchQuery ||
 			this.queryParams.dateModified ||
 			this.queryParams.dateCreated;
-	}
-
-	changeSort({ detail = {} }) {
-		const { sortKey, sortQuery } = detail;
-		if (/^(createdAt|updatedAt)$/.test(sortKey)) {
-			this.dateField = sortKey;
-		}
-
-		this._navigate('/my-videos', {
-			...this.queryParams,
-			sortQuery
-		});
 	}
 
 	contentListItemDeletedHandler(e) {
@@ -181,99 +97,6 @@ class ContentList extends DependencyRequester(InternalLocalizeMixin(NavigationMi
 		}
 	}
 
-	getCompareBasedOnSort(item) {
-		const itemUpdatedAtDate = new Date(item.updatedAt);
-		switch (this.queryParams.sortQuery) {
-			case 'updatedAt:desc':
-				return e => {
-					const elementUpdatedAtDate = new Date(e.updatedAt);
-					return elementUpdatedAtDate <= itemUpdatedAtDate;
-				};
-
-			case 'updatedAt:asc':
-				return e => {
-					const elementUpdatedAtDate = new Date(e.updatedAt);
-					return elementUpdatedAtDate >= itemUpdatedAtDate;
-				};
-
-			case 'lastRevTitle.keyword:desc':
-				return e => {
-					return e.title.toLowerCase() <= item.title.toLowerCase();
-				};
-
-			case 'lastRevTitle.keyword:asc':
-				return e => {
-					return e.title.toLowerCase() >= item.title.toLowerCase();
-				};
-
-			default:
-				return () => true;
-		}
-	}
-
-	async insertIntoContentItemsBasedOnSort(item) {
-		let indexToInsertAt = this._videos.findIndex(this.getCompareBasedOnSort(item));
-		if (indexToInsertAt === -1) {
-			indexToInsertAt = this._videos.length;
-		}
-
-		if (!this._moreResultsAvailable || indexToInsertAt !== this._videos.length) {
-			this._videos.splice(indexToInsertAt, 0, item);
-			this.requestUpdate();
-			await this.updateComplete;
-		}
-	}
-
-	async loadNext({ append = true } = {}) {
-		this.loading = true;
-		const { sortQuery, searchQuery, dateModified, dateCreated } = this.queryParams;
-		await this._handleVideoSearch({
-			append,
-			createdAt: dateCreated,
-			query: searchQuery,
-			sort: sortQuery,
-			updatedAt: dateModified,
-		});
-
-		this.loading = false;
-	}
-
-	observeQueryParams() {
-		observe(
-			rootStore.routingStore,
-			'queryParams',
-			change => {
-				if (this.loading) {
-					return;
-				}
-
-				const {
-					searchQuery: updatedSearchQuery = '',
-					sortQuery: updatedSortQuery = 'updatedAt:desc',
-					dateCreated: updatedDateCreated = '',
-					dateModified: updatedDateModified = ''
-				} = toJS(change.newValue);
-
-				const { searchQuery, sortQuery, dateCreated, dateModified } =
-					this.queryParams;
-
-				if (updatedSearchQuery === searchQuery && updatedSortQuery === sortQuery &&
-					updatedDateCreated === dateCreated && updatedDateModified === dateModified
-				) {
-					return;
-				}
-
-				this.queryParams = {
-					searchQuery: updatedSearchQuery,
-					sortQuery: updatedSortQuery,
-					dateCreated: updatedDateCreated,
-					dateModified: updatedDateModified
-				};
-				this.reloadPage();
-			}
-		);
-	}
-
 	observeSuccessfulUpload() {
 		observe(
 			this.uploader,
@@ -286,32 +109,6 @@ class ContentList extends DependencyRequester(InternalLocalizeMixin(NavigationMi
 				}
 			}
 		);
-	}
-
-	onWindowScroll() {
-		const contentListElem = this.shadowRoot.querySelector('#d2l-content-store-list');
-		if (contentListElem) {
-			const bottom = contentListElem.getBoundingClientRect().top + window.pageYOffset + contentListElem.clientHeight;
-			const scrollY = window.pageYOffset + window.innerHeight;
-			if (bottom - scrollY < this.infiniteScrollThreshold && this._moreResultsAvailable && !this.loading) {
-				this.loadNext();
-			}
-		}
-	}
-
-	async reloadPage() {
-		this.loading = true;
-		this._videos = [];
-		this._start = 0;
-		this._navigate('/my-videos', this.queryParams);
-
-		try {
-			await this.loadNext({ append: false });
-		} catch (error) {
-			this.loading = false;
-			this._videos = [];
-			this._start = 0;
-		}
 	}
 
 	renderContentItem(item) {
@@ -337,14 +134,6 @@ class ContentList extends DependencyRequester(InternalLocalizeMixin(NavigationMi
 				<content-list-item-ghost ?hidden=${!this.loading}></content-list-item-ghost>
 			</d2l-list>
 		`);
-	}
-
-	renderNotFound() {
-		return !this.loading && this._videos.length === 0 ? html`
-			<div class="d2l-capture-central-content-list-no-results">
-				${this.localize('noResults')}
-			</div>
-		` : html``;
 	}
 
 	showUndoDeleteToast() {
